refactor: initialized server in index
